Fix: Property type.

+++ L.GastosProdutos.Core/Domain/Entities/Packing/PackingEntity.cs
@@ -6,7 +6,7 @@ namespace L.GastosProdutos.Core.Domain.Entities.Packing
     {
         public PackingEntity() { }
 
-        public PackingEntity(string name, string? description, decimal price, float quantity)
+        public PackingEntity(string name, string? description, decimal price, decimal quantity)
         {
             Name = name;
             Description = description;
